+++ gvm/protocols/gmpv214/__init__.py
@@ -119,7 +119,6 @@ from gvm.protocols.gmpv208.entities.tickets import (
     get_ticket_status_from_string,
 )
 from gvm.protocols.gmpv208.entities.tls_certificates import TLSCertificateMixin
-from gvm.protocols.gmpv208.entities.trashcan import TrashcanMixin
 from gvm.protocols.gmpv208.entities.user_settings import UserSettingsMixin
 from gvm.protocols.gmpv208.entities.users import (
     UserAuthType,
@@ -127,6 +126,7 @@ from gvm.protocols.gmpv208.entities.users import (
 )
 from gvm.protocols.gmpv208.entities.vulnerabilities import VulnerabilitiesMixin
 
+from gvm.protocols.gmpv208.system.trashcan import TrashcanMixin
 
 from gvm.protocols.gmpv208.gmpv208 import GmpV208Mixin
