+++ tests/protocols/gmpv208/system/trashcan/test_empty_trashcan.py
