+++ tests/protocols/gmpv208/system/trashcan/test_restore_from_trashcan.py
